# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  email?: string | null;
  firstName?: string | null;
  lastName?: string | null;
  profileImageUrl?: string | null;
}

export interface GetCurrentAuthUserResponse {
  user: AuthUser | null;
}

export interface EmailRegisterRequest {
  email: string;
  /** @minLength 8 */
  password: string;
  firstName?: string;
  lastName?: string;
}

export interface EmailLoginRequest {
  email: string;
  password: string;
}

export interface ExchangeMobileAuthorizationCodeBody {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface ExchangeMobileAuthorizationCodeResponse {
  token: string;
}

export const LogoutMobileSessionResponseValue = {
  success: true,
} as const;
export type LogoutMobileSessionResponse =
  typeof LogoutMobileSessionResponseValue;

export interface ErrorEnvelope {
  error: string;
}

export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
  returnTo?: string;
};