# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Get the currently authenticated user
 */
export const GetCurrentAuthUserHeader = zod.object({
  Authorization: zod.string().optional(),
});

export const GetCurrentAuthUserResponse = zod.object({
  user: zod.union([
    zod.object({
      id: zod.string(),
      email: zod.string().nullish(),
      firstName: zod.string().nullish(),
      lastName: zod.string().nullish(),
      profileImageUrl: zod.string().nullish(),
    }),
    zod.null(),
  ]),
});

/**
 * @summary Register a new account with email and password
 */
export const registerWithEmailBodyPasswordMin = 8;

export const RegisterWithEmailBody = zod.object({
  email: zod.string().email(),
  password: zod.string().min(registerWithEmailBodyPasswordMin),
  firstName: zod.string().optional(),
  lastName: zod.string().optional(),
});

/**
 * @summary Login with email and password
 */
export const LoginWithEmailBody = zod.object({
  email: zod.string().email(),
  password: zod.string(),
});

export const LoginWithEmailResponse = zod.object({
  user: zod.union([
    zod.object({
      id: zod.string(),
      email: zod.string().nullish(),
      firstName: zod.string().nullish(),
      lastName: zod.string().nullish(),
      profileImageUrl: zod.string().nullish(),
    }),
    zod.null(),
  ]),
});

/**
 * @summary Start the browser OIDC login flow
 */
export const BeginBrowserLoginQueryParams = zod.object({
  returnTo: zod.coerce.string().optional(),
});

/**
 * @summary Exchange a mobile OIDC code for a session token
 */

export const ExchangeMobileAuthorizationCodeBody = zod.object({
  code: zod.string().min(1),
  code_verifier: zod.string().min(1),
  redirect_uri: zod.string().url().min(1),
  state: zod.string().min(1),
  nonce: zod.string().min(1).optional(),
});

export const ExchangeMobileAuthorizationCodeResponse = zod.object({
  token: zod.string(),
});

/**
 * @summary Delete a mobile session token
 */
export const LogoutMobileSessionResponse = zod.object({
  success: zod.boolean(),
});