# lib/api-zod/src/generated/types/emailRegisterRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface EmailRegisterRequest {
  email: string;
  /** @minLength 8 */
  password: string;
  firstName?: string;
  lastName?: string;
}